services/ai-worker/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
psycopg2-binary>=2.9
google-genai>=1.0
pydantic>=2.0

# Dev/Test
pytest>=8.0
httpx>=0.27
